% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_build.R
\name{dwv_add_student}
\alias{dwv_add_student}
\title{Give students r/w access to their repository}
\usage{
dwv_add_student(x)
}
\description{
Give students r/w access to their repository
}
\note{
https://developer.github.com/v3/repos/collaborators/#add-user-as-a-collaborator
}
