% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_build.R
\name{dwv_delete_github}
\alias{dwv_delete_github}
\title{Github Repo Delete}
\usage{
dwv_delete_github(repo_name, owner_name = "BYUI335")
}
\arguments{
\item{repo_name}{is the name of the repo to delete.}

\item{owner_name}{is the Github group or user where the package is stored.}
}
\description{
Github Repo Delete
}
