% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_build.R
\name{dwv_invite_group}
\alias{dwv_invite_group}
\title{Invite missing to the team.  They should get an email.}
\usage{
dwv_invite_group(x, students_team_id)
}
\description{
Invite missing to the team.  They should get an email.
}
