% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_issues.R
\name{dwv_issue_tasks}
\alias{dwv_issue_tasks}
\title{Create task issues}
\usage{
dwv_issue_tasks(
  x = body_lines,
  task_title = "Daily Class Tasks",
  dat = repo_create_df_new
)
}
\description{
Create task issues
}
