% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwv_one_student.R
\name{dwv_one_student}
\alias{dwv_one_student}
\title{Function to create repo and connect them to group plus push issues}
\usage{
dwv_one_student(one_student, class_group, ta_group, body_lines = NULL)
}
\arguments{
\item{class_group}{the name of the class group for that semester.}

\item{ta_group}{the name of the TA group for that semester.}

\item{body_lines}{is the MarkDown file with a predefined format}

\item{one_stdent}{is a tibble with all the information for one person}
}
\description{
Function to create repo and connect them to group plus push issues
}
