% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwv_one_student.R
\name{dwv_one_student_issues}
\alias{dwv_one_student_issues}
\title{Function to create issues}
\usage{
dwv_one_student_issues(one_student, body_lines)
}
\arguments{
\item{body_lines}{is the MarkDown file with a predefined format}

\item{one_stdent}{is a tibble with all the information for one person}
}
\description{
Function to create issues
}
