% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_build.R
\name{ghb_invite_org}
\alias{ghb_invite_org}
\title{Invite into Organization}
\usage{
ghb_invite_org(email, org, team_ids = NULL, role = "direct_member")
}
\arguments{
\item{email}{is the email allowed in org}

\item{org}{is the created org}

\item{team_ids}{are the team ids returned from github}

\item{role}{can be "admin" or "direct_member" defaults to "direct_member"}
}
\value{
A table with "name", "value", and "type" columns as digested from the API.
}
\description{
Invite members into the org
}
\details{
see "https://docs.github.com/en/enterprise-cloud@latest/rest/orgs/members?apiVersion=2022-11-28#create-an-organization-invitation". The id in the table is the invitation id.
}
