% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_pull.R
\name{git_clone}
\alias{git_clone}
\title{Clone Repo}
\usage{
git_clone(git_dir, group_dir, repo_use)
}
\arguments{
\item{git_dir}{is the folder on your local computer where you store your git repositories}

\item{group_dir}{is the folder for the GitHub group id.  It}

\item{repo_use}{is the repository name for the clone}
}
\description{
Clone Repo
}
