% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_pull.R
\name{git_pull}
\alias{git_pull}
\title{Pull Latest from Repo}
\usage{
git_pull(git_dir, group_dir, repo_use)
}
\arguments{
\item{git_dir}{is the folder on your local computer where you store your git repositories}

\item{group_dir}{is the folder for the GitHub group id.}

\item{repo_use}{is the repository name for the pull}
}
\description{
Pull Latest from Repo
}
