% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_pull.R
\name{git_push}
\alias{git_push}
\title{Commit and Push Latest Updates to Repo}
\usage{
git_push(git_dir, group_dir, repo_use, message = "'First Push from Hathaway'")
}
\arguments{
\item{git_dir}{is the folder on your local computer where you store your git repositories}

\item{group_dir}{is the folder for the GitHub group id.}

\item{repo_use}{is the repository name for the push}

\item{message}{is the commit message to use}
}
\description{
Commit and Push Latest Updates to Repo
}
