% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_pull.R
\name{list_repos}
\alias{list_repos}
\title{List all Repos}
\usage{
list_repos(group_dir, org = TRUE)
}
\arguments{
\item{group_dir}{is the folder for the GitHub group id.}
}
\value{
a vector of repository names.  Can be used in `git_clone` and `git_pull`
}
\description{
List all Repos
}
